requirements.txt
Flask
python-dotenv
rich>=11.2.0
requests
python-dateutil 